requirements.txt
# for test
pytest
pytest-django
pytest-cov
django-test-plus
model_mommy